# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DYL Real-Estate Services API
  description: >
    Admin-verified real estate listings API (Lagos-first, Nigeria). Three
    roles share this API: **seekers** (browse/save/inquire, the default
    `role` on every account), **owners** (list/manage their own
    properties — a seeker becomes an owner the moment they create a
    listing), and **admins** (`role: admin`) who review and verify
    listings and moderate the platform.

    Authentication is a bearer JWT returned by `/auth/login`,
    `/auth/register`, or `/auth/google`. Send it as
    `Authorization: Bearer <token>` on every authenticated request.
    Tokens are invalidated (not just expired) when the account's
    password changes — see `token_version` under Auth below.
  version: "1.0.0"
servers:
  - url: /api
    description: Relative to wherever this API is deployed (see README for
      environment-specific base URLs).
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from `/auth/login`, `/auth/register`, or `/auth/google`.
        Carries `id`, `role`, `email`, `name`, and `token_version` — the
        last of which is checked against the user's current value on
        every request, so it's what lets a password change or reset
        immediately invalidate every previously-issued token rather than
        waiting out the token's natural expiry.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: 1-indexed page number.
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 47 }
        totalPages: { type: integer, example: 3 }
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: Search public listings
      description: Always excludes `status != active` and
        `verification_status = rejected` — a rejected listing is never
        publicly visible regardless of other filters. Page size caps at 50.
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Free-text search across title, description, city, state.
        - name: state
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: property_type
          in: query
          schema: { type: string, enum: [land, residential, commercial] }
        - name: listing_type
          in: query
          schema: { type: string, enum: [sale, rent] }
        - name: min_price
          in: query
          schema: { type: number }
        - name: max_price
          in: query
          schema: { type: number }
        - name: verified_only
          in: query
          schema: { type: boolean }
        - name: has_video
          in: query
          schema: { type: boolean }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200":
          description: OK — possibly an empty `data` array; never an error
            for "no matches".
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
    post:
      tags: [Properties]
      summary: Create a listing
      description: Owned by the authenticated caller; always starts
        `verification_status = pending` regardless of what's sent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Property" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /posts:
    get:
      tags: [Posts]
      summary: List published blog posts
      description: Only ever returns `published = true` posts — a draft's
        slug 404s if hit directly (see `/posts/{slug}`).
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Post" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
  /posts/{slug}:
    get:
      tags: [Posts]
      summary: Get one published post
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Post" } }
        "404":
          description: Doesn't exist, or exists but isn't published.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/stats:
    get:
      tags: [Admin]
      summary: Dashboard summary numbers
      responses:
        "200": { description: OK. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/properties:
    get:
      tags: [Admin]
      summary: List all listings (any status)
      description: Unlike public search, includes pending/rejected/sold —
        this is the review queue. Max page size 100.
      parameters:
        - name: verification_status
          in: query
          schema: { type: string, enum: [pending, verified, rejected] }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/properties/{id}/verify:
    patch:
      tags: [Admin]
      summary: Verify, reject, or reset a listing's review status
      description: Emails the owner (best-effort) unless resetting to
        `pending`, which is a neutral internal state change with nothing
        actionable for them. Logged to the audit trail either way.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verification_status]
              properties:
                verification_status: { type: string, enum: [verified, rejected, pending] }
                verification_notes:
                  type: string
                  description: Shown to the owner in their rejection email
                    when status is `rejected`; internal otherwise.
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Property" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      description: Never includes `password_hash`. Max page size 100.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/User" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Promote or demote a user
      description: An admin cannot change their own role this way (400) —
        prevents an admin locking themselves, and everyone else, out of
        the console with no way back short of a direct database edit.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties: { role: { type: string, enum: [user, admin] } }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "400":
          description: Validation failed, or attempting to change your
            own role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a user
      description: An admin cannot delete their own account this way
        (400). Blocked (409) if the user still owns any listings —
        delete or reassign those first.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted. }
        "400":
          description: Attempting to delete your own account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: User still owns listings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/inquiries:
    get:
      tags: [Admin]
      summary: List inquiries
      parameters:
        - name: property_id
          in: query
          schema: { type: integer }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Inquiry" } }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...